requirements.txt
python-dotenv>=1.0.1
pyYAML>=6.0.1
blue-yonder>=0.1.11
pillow>=11.1.0